requirements.txt
aniso8601==7.0.0
asgiref==3.5.2
autopep8==1.7.1
backports.zoneinfo==0.2.1
black==22.10.0
click==8.1.3
Django==3.2.16
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==2.4.0
django-graphql-auth==0.3.15
django-graphql-jwt==0.3.0
djangorestframework==3.14.0
graphene==2.1.8
graphene-django==2.13.0
graphene-file-upload==1.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
mypy-extensions==0.4.3
pathspec==0.10.1
Pillow==9.3.0
platformdirs==2.5.2
promise==2.3
pycodestyle==2.9.1
PyJWT==1.7.1
pytz==2020.4
Rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.4.1
tomli==2.0.1
typing-extensions==4.4.0
Unidecode==1.1.1
psycopg2-binary==2.9.5
dj-database-url==1.0.0
gunicorn==20.1.0